Move file IO methods to separate class
Motivation: method readEdgeList is used in RunNetworkClustering and RunNetworkLayout
