+++ src/cwts/networkanalysis/run/RunNetworkClustering.java
@@ -6,16 +6,6 @@ import cwts.networkanalysis.IterativeCPMClusteringAlgorithm;
 import cwts.networkanalysis.LeidenAlgorithm;
 import cwts.networkanalysis.LouvainAlgorithm;
 import cwts.networkanalysis.Network;
-import cwts.util.DynamicDoubleArray;
-import cwts.util.DynamicIntArray;
-import java.io.BufferedReader;
-import java.io.BufferedWriter;
-import java.io.FileNotFoundException;
-import java.io.FileReader;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.io.OutputStreamWriter;
-import java.util.Arrays;
 import java.util.Random;
 
 /**
@@ -141,11 +131,6 @@ public final class RunNetworkClustering
           + "-o --output-clustering <filename> (default: standard output)\n"
           + "    Write the final clustering to the specified file. If no file is specified,\n"
           + "    the standard output is used.\n";
-
-    /**
-     * Column separator for edge list and clustering files.
-     */
-    public static final String COLUMN_SEPARATOR = "\t";
 
     /**
      * This method is called when the tool is started.
@@ -327,7 +312,7 @@ public final class RunNetworkClustering
         // Read edge list from file.
         System.err.println("Reading " + (sortedEdgeList ? "sorted " : "") + "edge list from '" + edgeListFilename + "'.");
         long startTimeEdgeListFile = System.currentTimeMillis();
-        Network network = readEdgeList(edgeListFilename, weightedEdges, sortedEdgeList);
+        Network network = FileIO.readEdgeList(edgeListFilename, weightedEdges, sortedEdgeList);
         System.err.println("Reading " + (sortedEdgeList ? "sorted " : "") + "edge list took " + (System.currentTimeMillis() - startTimeEdgeListFile) / 1000 + "s.");
         System.err.println("Network consists of " + network.getNNodes() + " nodes and " + network.getNEdges() + " edges" + (weightedEdges ? " with a total edge weight of " + network.getTotalEdgeWeight() : "") + ".");
 
@@ -336,7 +321,7 @@ public final class RunNetworkClustering
         if (initialClusteringFilename != null)
         {
             System.err.println("Reading initial clustering from '" + initialClusteringFilename + "'.");
-            initialClustering = readClustering(initialClusteringFilename, network.getNNodes());
+            initialClustering = FileIO.readClustering(initialClusteringFilename, network.getNNodes());
             System.err.println("Initial clustering consists of " + initialClustering.getNClusters() + " clusters.");
         }
         else
@@ -391,240 +376,7 @@ public final class RunNetworkClustering
 
         // Write final clustering to file (or to standard output).
         System.err.println("Writing final clustering to " + ((finalClusteringFilename == null) ? "standard output." : "'" + finalClusteringFilename + "'."));
-        writeClustering(finalClusteringFilename, finalClustering);
-    }
-
-    /**
-     * Reads an edge list from a file and creates a network.
-     *
-     * @param filename       Filename
-     * @param weightedEdges  Indicates whether edges have weights
-     * @param sortedEdgeList Indicates whether the edge list is sorted
-     *
-     * @return Network
-     */
-    public static Network readEdgeList(String filename, boolean weightedEdges, boolean sortedEdgeList)
-    {
-        // Read edge list.
-        DynamicIntArray[] edges = new DynamicIntArray[2];
-        edges[0] = new DynamicIntArray(100);
-        edges[1] = new DynamicIntArray(100);
-        DynamicDoubleArray edgeWeights = weightedEdges ? new DynamicDoubleArray(100) : null;
-        int nNodes = 0;
-        BufferedReader reader = null;
-        try
-        {
-            reader = new BufferedReader(new FileReader(filename));
-            String line = reader.readLine();
-            int lineNo = 0;
-            while (line != null)
-            {
-                lineNo++;
-                String[] columns = line.split(COLUMN_SEPARATOR);
-                if ((!weightedEdges && ((columns.length < 2) || (columns.length > 3))) || (weightedEdges && (columns.length != 3)))
-                    throw new IOException("Incorrect number of columns (line " + lineNo + ").");
-
-                int node1;
-                int node2;
-                try
-                {
-                    node1 = Integer.parseUnsignedInt(columns[0]);
-                    node2 = Integer.parseUnsignedInt(columns[1]);
-                }
-                catch (NumberFormatException e)
-                {
-                    throw new IOException("Node must be represented by a zero-index integer number (line " + lineNo + ").");
-                }
-                edges[0].append(node1);
-                edges[1].append(node2);
-                if (node1 >= nNodes)
-                    nNodes = node1 + 1;
-                if (node2 >= nNodes)
-                    nNodes = node2 + 1;
-
-                if (weightedEdges)
-                {
-                    double weight;
-                    try
-                    {
-                        weight = Double.parseDouble(columns[2]);
-                    }
-                    catch (NumberFormatException e)
-                    {
-                        throw new IOException("Edge weight must be a number (line " + lineNo + ").");
-                    }
-                    edgeWeights.append(weight);
-                }
-
-                line = reader.readLine();
-            }
-        }
-        catch (FileNotFoundException e)
-        {
-            System.err.println("Error while reading edge list from file: File not found.");
-            System.exit(-1);
-        }
-        catch (IOException e)
-        {
-            System.err.println("Error while reading edge list from file: " + e.getMessage());
-            System.exit(-1);
-        }
-        finally
-        {
-            if (reader != null)
-                try
-                {
-                    reader.close();
-                }
-                catch (IOException e)
-                {
-                    System.err.println("Error while reading edge list from file: " + e.getMessage());
-                    System.exit(-1);
-                }
-        }
-
-        // Create network.
-        Network network = null;
-        int[][] edges2 = new int[2][];
-        edges2[0] = edges[0].toArray();
-        edges2[1] = edges[1].toArray();
-        try
-        {
-            if (weightedEdges)
-                network = new Network(nNodes, true, edges2, edgeWeights.toArray(), sortedEdgeList, true);
-            else
-                network = new Network(nNodes, true, edges2, sortedEdgeList, true);
-        }
-        catch (IllegalArgumentException e)
-        {
-            System.err.println("Error while creating network: " + e.getMessage());
-            System.exit(-1);
-        }
-        return network;
-    }
-
-    /**
-     * Reads a clustering from a file.
-     *
-     * @param filename Filename
-     * @param nNodes   Number of nodes
-     *
-     * @return Clustering
-     */
-    public static Clustering readClustering(String filename, int nNodes)
-    {
-        int[] clusters = new int[nNodes];
-        Arrays.fill(clusters, -1);
-        BufferedReader reader = null;
-        try
-        {
-            reader = new BufferedReader(new FileReader(filename));
-            String line = reader.readLine();
-            int lineNo = 0;
-            while (line != null)
-            {
-                lineNo++;
-                String[] columns = line.split(COLUMN_SEPARATOR);
-                if (columns.length != 2)
-                    throw new IOException("Incorrect number of columns (line " + lineNo + ").");
-
-                int node;
-                try
-                {
-                    node = Integer.parseUnsignedInt(columns[0]);
-                }
-                catch (NumberFormatException e)
-                {
-                    throw new IOException("Node must be represented by a zero-index integer number (line " + lineNo + ").");
-                }
-                if (node >= nNodes)
-                    throw new IOException("Invalid node (line " + lineNo + ").");
-                int cluster;
-                try
-                {
-                    cluster = Integer.parseUnsignedInt(columns[1]);
-                }
-                catch (NumberFormatException e)
-                {
-                    throw new IOException("Cluster must be represented by a zero-index integer number (line " + lineNo + ").");
-                }
-                if (clusters[node] >= 0)
-                    throw new IOException("Duplicate node (line " + lineNo + ").");
-                clusters[node] = cluster;
-
-                line = reader.readLine();
-            }
-            if (lineNo < nNodes)
-                throw new IOException("Missing nodes.");
-        }
-        catch (FileNotFoundException e)
-        {
-            System.err.println("Error while reading clustering from file: File not found.");
-            System.exit(-1);
-        }
-        catch (IOException e)
-        {
-            System.err.println("Error while reading clustering from file: " + e.getMessage());
-            System.exit(-1);
-        }
-        finally
-        {
-            if (reader != null)
-                try
-                {
-                    reader.close();
-                }
-                catch (IOException e)
-                {
-                    System.err.println("Error while reading clustering from file: " + e.getMessage());
-                    System.exit(-1);
-                }
-        }
-
-        return new Clustering(clusters);
-    }
-
-    /**
-     * Writes a clustering to a file.
-     *
-     * @param filename   Filename
-     * @param clustering Clustering
-     */
-    public static void writeClustering(String filename, Clustering clustering)
-    {
-        BufferedWriter writer = null;
-        try
-        {
-            writer = new BufferedWriter((filename == null) ? new OutputStreamWriter(System.out) : new FileWriter(filename));
-            for (int i = 0; i < clustering.getNNodes(); i++)
-            {
-                writer.write(i + COLUMN_SEPARATOR + clustering.getCluster(i));
-                writer.newLine();
-            }
-        }
-        catch (FileNotFoundException e)
-        {
-            System.err.println("Error while writing clustering to file: File not found.");
-            System.exit(-1);
-        }
-        catch (IOException e)
-        {
-            System.err.println("Error while writing clustering to file: " + e.getMessage());
-            System.exit(-1);
-        }
-        finally
-        {
-            if (writer != null)
-                try
-                {
-                    writer.close();
-                }
-                catch (IOException e)
-                {
-                    System.err.println("Error while writing clustering to file: " + e.getMessage());
-                    System.exit(-1);
-                }
-        }
+        FileIO.writeClustering(finalClusteringFilename, finalClustering);
     }
 
     private RunNetworkClustering()
